cards v0.1.0 Cards

Provides methods for creating and handling a deck of cards.

Functions

create_deck()

Returns a list of strings representing a deck of playing cards.

create_hand(cards_number)

hand(deck, hand_size)

Devides a deck into a hand and a remainder of the deck. The hand_size argument indicates how many cards should be in hand

Examples

iex> deck = Cards.create_deck
iex> {hand, deck} = Cards.hand(deck, 1)
iex> hand
["Ace of Spides"]

include?(deck, card)

load(filename)

save(deck, filename)

shuffle_deck(deck)